// src/app/shared/model/jerarquia.ts
export interface Jerarquia {
  codigo: number;
  nombre: string;
  nivel: number;
  codigoPadre: number;
  descripcion: string;
  ultimoNivel: number;
  rubroTipoEstructuraP: number;
  rubroTipoEstructuraH: number;
  codigoAlterno: number;
  rubroNivelCaracteristicaP: number;
  rubroNivelCaracteristicaH: number;
}